refactor: main.basic.js 내 일부함수 utils로 이동
