+++ src/basic/main.basic.js
@@ -1,11 +1,7 @@
-import { updateBonusPoints, updateTotalInfo } from './utils/cart';
-import {
-  calculateDayDiscount,
-  calculateDiscountedPrice,
-  calculateTotalProductsBulkDiscount,
-  getProductBulkDiscountRate,
-} from './utils/discount';
-import { products, renderProductsStockInfo } from './utils/product';
+import { updateCartInfos } from './utils/cart';
+import { addToCart, changeCartItemQuantity, removeCartItem } from './utils/cartOperations';
+import { setSuggestDiscount, setSurpriseDiscount } from './utils/discount';
+import { products, renderProductOptions } from './utils/product';
 
 let productSelectDropdown, addToCartBtn, cartItemsDisplay, cartTotalInfo, productsStockInfo;
 let lastAddedProduct,
@@ -14,9 +10,10 @@ let lastAddedProduct,
 const main = () => {
   renderCartUI();
   renderProductOptions();
+  bonusPoints = updateCartInfos(bonusPoints);
 
-  updateCartInfos();
-  scheduleRandomSales();
+  scheduleRandomDiscount();
+  setEventListeners();
 };
 
 const renderCartUI = () => {
@@ -56,122 +53,9 @@ const renderCartUI = () => {
   cartWrapper.appendChild(cartContainer);
   root.appendChild(cartWrapper);
 };
-const renderProductOptions = () => {
-  const productSelectDropdown = document.getElementById('product-select');
-  productSelectDropdown.innerHTML = '';
 
-  products.forEach((item) => {
-    const opt = document.createElement('option');
-    opt.value = item.id;
-
-    opt.textContent = item.name + ' - ' + item.price + '원';
-    if (item.quantity === 0) opt.disabled = true;
-    productSelectDropdown.appendChild(opt);
-  });
-};
-const updateCartInfos = () => {
-  let totalItems = 0;
-  let totalPrice = 0;
-  let discountedTotalPrice = 0;
-
-  const cartItems = cartItemsDisplay.children;
-  for (let cartItem of cartItems) {
-    const currentProduct = products.find((product) => product.id === cartItem.id);
-    const quantity = parseInt(cartItem.querySelector('span').textContent.split('x ')[1]);
-    const productTotalPrice = currentProduct.price * quantity;
-
-    totalItems += quantity;
-    totalPrice += productTotalPrice;
-
-    const productBulkDiscountRate = getProductBulkDiscountRate(currentProduct.id, quantity);
-    discountedTotalPrice += calculateDiscountedPrice(productTotalPrice, productBulkDiscountRate);
-  }
-
-  const updatedTotalPriceAndDiscountRate = calculateTotalProductsBulkDiscount(
-    totalItems,
-    totalPrice,
-    discountedTotalPrice,
-  );
-  const { updatedTotalPrice, discountRate } = calculateDayDiscount(updatedTotalPriceAndDiscountRate);
-
-  updateTotalInfo(updatedTotalPrice, discountRate);
-  bonusPoints = updateBonusPoints(bonusPoints, updatedTotalPrice);
-  renderProductsStockInfo();
-};
-const scheduleRandomSales = () => {
-  setTimeout(() => {
-    setInterval(() => {
-      const luckyItem = products[Math.floor(Math.random() * products.length)];
-      if (Math.random() < 0.3 && luckyItem.quantity > 0) {
-        luckyItem.price = Math.round(luckyItem.price * 0.8);
-        alert('번개세일! ' + luckyItem.name + '이(가) 20% 할인 중입니다!');
-        renderProductOptions();
-      }
-    }, 30000);
-  }, Math.random() * 10000);
-
-  setTimeout(() => {
-    setInterval(() => {
-      if (lastAddedProduct) {
-        const suggest = products.find((product) => {
-          return product.id !== lastAddedProduct && product.quantity > 0;
-        });
-        if (suggest) {
-          alert(suggest.name + '은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!');
-          suggest.price = Math.round(suggest.price * 0.95);
-          renderProductOptions();
-        }
-      }
-    }, 60000);
-  }, Math.random() * 20000);
-};
-
-main();
-
-const createNewItem = (targetProduct) => {
-  const newItem = document.createElement('div');
-  newItem.id = targetProduct.id;
-  newItem.className = 'flex justify-between items-center mb-2';
-  newItem.innerHTML =
-    '<span>' +
-    targetProduct.name +
-    ' - ' +
-    targetProduct.price +
-    '원 x 1</span><div>' +
-    '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
-    targetProduct.id +
-    '" data-change="-1">-</button>' +
-    '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
-    targetProduct.id +
-    '" data-change="1">+</button>' +
-    '<button class="remove-item bg-red-500 text-white px-2 py-1 rounded" data-product-id="' +
-    targetProduct.id +
-    '">삭제</button></div>';
-  cartItemsDisplay.appendChild(newItem);
-};
-const addToCart = (targetCartItem, targetProduct) => {
-  if (!targetCartItem) {
-    createNewItem(targetProduct);
-    targetProduct.quantity--;
-    return;
-  }
-
-  const currentItemQuantity = parseInt(targetCartItem.querySelector('span').textContent.split('x ')[1]);
-  const newItemQuantity = currentItemQuantity + 1;
-
-  // BUG: 로직 개선
-  const isStockRemain = newItemQuantity <= targetProduct.quantity;
-  if (isStockRemain) {
-    const newCartItemInfo = `${targetProduct.name} - ${targetProduct.price}원 x ${newItemQuantity}`;
-
-    targetCartItem.querySelector('span').textContent = newCartItemInfo;
-    targetProduct.quantity--;
-  } else {
-    alert('재고가 부족합니다.');
-  }
-};
 const handleAddToCart = () => {
-  const selectedProductId = productSelectDropdown.value;
+  const selectedProductId = document.getElementById('product-select').value;
   const targetProduct = products.find((product) => {
     return product.id === selectedProductId;
   });
@@ -180,32 +64,8 @@ const handleAddToCart = () => {
     const targetCartItem = document.getElementById(targetProduct.id);
 
     addToCart(targetCartItem, targetProduct);
-    updateCartInfos();
+    bonusPoints = updateCartInfos(bonusPoints);
     lastAddedProduct = selectedProductId;
-  }
-};
-addToCartBtn.addEventListener('click', handleAddToCart);
-
-// TODO: handleAddToCart랑 중복되는 부분 합칠 수 있는 지 확인
-const removeCartItem = (targetProduct, itemElement, restoreQuantity) => {
-  targetProduct.quantity += restoreQuantity;
-  itemElement.remove();
-};
-const changeCartItemQuantity = (clickedElement, itemElement, targetProduct) => {
-  const quantityChangeAmount = parseInt(clickedElement.dataset.change);
-  const currentItemQuantity = parseInt(itemElement.querySelector('span').textContent.split('x ')[1]);
-  const newItemQuantity = currentItemQuantity + quantityChangeAmount;
-
-  const isStockRemain = newItemQuantity <= targetProduct.quantity + currentItemQuantity;
-
-  if (!isStockRemain) {
-    alert('재고가 부족합니다.');
-  } else if (newItemQuantity > 0) {
-    itemElement.querySelector('span').textContent =
-      itemElement.querySelector('span').textContent.split('x ')[0] + 'x ' + newItemQuantity;
-    targetProduct.quantity -= quantityChangeAmount;
-  } else {
-    removeCartItem(targetProduct, itemElement, 1);
   }
 };
 const handleCartItemsDisplay = (event) => {
@@ -228,6 +88,16 @@ const handleCartItemsDisplay = (event) => {
     removeCartItem(targetProduct, itemElement, currentItemQuantity);
   }
 
-  updateCartInfos();
+  bonusPoints = updateCartInfos(bonusPoints);
 };
-cartItemsDisplay.addEventListener('click', handleCartItemsDisplay);
+const setEventListeners = () => {
+  addToCartBtn.addEventListener('click', handleAddToCart);
+  cartItemsDisplay.addEventListener('click', handleCartItemsDisplay);
+};
+
+const scheduleRandomDiscount = () => {
+  setSurpriseDiscount();
+  setSuggestDiscount(lastAddedProduct);
+};
+
+main();
