+++ src/basic/utils/discount.js
@@ -1,3 +1,41 @@
+import { products, renderProductOptions } from './product';
+
+export const setSurpriseDiscount = () => {
+  const SURPRISE_DISCOUNT_PROBABILITY = 0.3;
+  const SURPRISE_DISCOUNT_RATE = 0.8;
+  const SURPRISE_TIME_INTERVAL = 30000;
+
+  setTimeout(() => {
+    setInterval(() => {
+      const luckyItem = products[Math.floor(Math.random() * products.length)];
+      if (Math.random() < SURPRISE_DISCOUNT_PROBABILITY && luckyItem.quantity > 0) {
+        luckyItem.price = Math.round(luckyItem.price * SURPRISE_DISCOUNT_RATE);
+        alert(`번개세일! ${luckyItem.name}이(가) 20% 할인 중입니다!`);
+        renderProductOptions();
+      }
+    }, SURPRISE_TIME_INTERVAL);
+  }, Math.random() * 10000);
+};
+export const setSuggestDiscount = (lastAddedProduct) => {
+  const SUGGEST_DISCOUNT_RATE = 0.95;
+  const SUGGEST_TIME_INTERVAL = 60000;
+
+  setTimeout(() => {
+    setInterval(() => {
+      if (lastAddedProduct) {
+        const suggest = products.find((product) => {
+          return product.id !== lastAddedProduct && product.quantity > 0;
+        });
+        if (suggest) {
+          alert(`${suggest.name}은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!`);
+          suggest.price = Math.round(suggest.price * SUGGEST_DISCOUNT_RATE);
+          renderProductOptions();
+        }
+      }
+    }, SUGGEST_TIME_INTERVAL);
+  }, Math.random() * 20000);
+};
+
 const PRODUCT_BULK_DISCOUNT_AMOUNT = 10;
 const PRODUCT_BULK_DISCOUNT_RATE = {
   p1: 0.1,
